Update and rename 19_MakesBeepSoundWhenFaceDetected.py to 3_MakesBeepSoundWhenFaceDetected.py
"Ever wanted a program that spots faces in pictures or videos? That's exactly what our GitHub project does! And get this: whenever it finds a face, it beeps to let you know. It's perfect if you're just starting out and want to learn about face detection. We've made sure the instructions are super simple to follow. So whether you're into making security gadgets, cool cameras, or fun interactive stuff, you'll love playing with this project!"